Add decorators to class methods
