--- src/pizza/__init__.py
@@ -1,1 +1,0 @@
-from .cli import cli
