+++ src/pizza/business.py
@@ -1,99 +1,86 @@
 from collections import defaultdict
-from collections import UserDict
-from typing import Self
+import functools
+import time
+import random
+import os
+from pizza.pizza_menu import DictAssortment, FoodItem, PizzaRecipe
+
+from pizza.pizza_menu import assortment, Pepperoni
+
+LATENCY_ENABLED = os.getenv(
+    "LATENCY_ENABLED", "0"
+)  # latency added only for cli app
 
 
-class FoodItem:
-    """Super class for all food items"""
+def add_latency(fn):
+    @functools.wraps(fn)  # not sure if it works with class methods
+    def wrapper(self, *args, **kwargs):
+        # global LATENCY_ENABLED
+        time_sleep = 0
+        if LATENCY_ENABLED == "1":
+            time_sleep = random.randint(1, 3) / 5
+        time.sleep(time_sleep)
+        result = fn(self, *args, **kwargs)
+        return result
 
-    recipe: dict
-    emoji: str
-    type_of_food: str
+    return wrapper
 
 
-class PizzaRecipe(FoodItem):
-    """Specialization of food - pizza"""
+def log(str_template: str):
+    def outer_wrapper(fn):
+        @functools.wraps(fn)
+        def wrapper(self, *args, **kwargs):
+            time_start = time.time()
+            result = fn(self, *args, **kwargs)
+            lapsed_time = time.time() - time_start
+            print(str_template.format(lapsed_time))  # apply to placeholder
+            return result
 
-    type_of_food = "pizza"
+        return wrapper
 
-    def __init__(self, size="L") -> None:
-        self.size = size
-        self.is_baked = False
-
-    def bake(self):
-        """Might be incorrect"""
-        if not self.is_baked:
-            self.is_baked = True
-
-    @classmethod
-    @property
-    def name(cls) -> str:
-        return cls.__mro__[0].__name__.capitalize()
-
-    @classmethod
-    @property
-    def clean_recipe(cls) -> str:
-        """Dict of ingredients to a string"""
-        return ", ".join([i for i in cls.recipe])
-
-    def __str__(self) -> str:
-        return (
-            f"{self.__class__.__name__} {self.type_of_food}, "
-            f"Size: {self.size}, Is baked: {self.is_baked}"
-        )
-
-    def __repr__(self) -> str:
-        return self.__class__.__name__
-
-    def __eq__(self, other: Self) -> bool:
-        return (
-            (set(self.recipe) == set(other.recipe))
-            & (self.name == other.name)
-            & (self.size == other.size)
-        )
-
-    def dict(self):
-        print(self.recipe)
+    return outer_wrapper
 
 
 class Restaurant:
     def __init__(self, menu: "DictAssortment") -> None:
         self.menu = menu
-        self.stock = defaultdict(list)
+        self._stock = defaultdict(list)
 
-    @staticmethod
-    def bake(pizza: PizzaRecipe) -> PizzaRecipe:
+    @log("Baking took {:.2f} seconds")
+    @add_latency
+    def bake(self, pizza: PizzaRecipe) -> PizzaRecipe:
         if not pizza.is_baked:
-            # add time sleep
-            print("baked")
+            print("👩‍🍳 baked", end=" ")
             pizza.is_baked = True
         return pizza
 
-    def add_to_stock(self, client: "Client", item: FoodItem) -> None:
-        self.stock[client].append(item)
+    def _add_to_stock(self, client: "Client", item: FoodItem) -> None:
+        self._stock[client].append(item)
 
-    def retrieve_from_stock(self, client: "Client") -> list[FoodItem]:
-        items = self.stock[client]
-        del self.stock[client]
+    def _retrieve_from_stock(self, client: "Client") -> list[FoodItem]:
+        items = self._stock[client]
+        del self._stock[client]
         return items
 
-    def pickup(self, client):
-        food = self.retrieve_from_stock(client)
-        print("picked up from restaurant")
+    def pickup(self, client: "Client") -> list[FoodItem]:
+        food = self._retrieve_from_stock(client)
         return food
 
-    def make_order(self, pizza_name, client: "Client", is_delivery=False):
-        print("was made a request for pizza")
+    def make_order(
+        self, pizza_name, client: "Client", is_delivery=False
+    ) -> None:
         pizza = self.menu[pizza_name]()
         pizza = self.bake(pizza)
-        self.add_to_stock(client, pizza)
+        self._add_to_stock(client, pizza)
         if is_delivery:
             self.deliver(client)
 
+    @log("Delivery took {:.2f} seconds")
+    @add_latency
     def deliver(self, client: "Client") -> None:
         # do some work with timer
-        print("delivered")
-        client.add_to_stock(self.retrieve_from_stock(client))
+        print("🛵 delivered", end=" ")
+        client.add_to_stock(self._retrieve_from_stock(client))
 
 
 class Client:
@@ -111,10 +98,10 @@ class Client:
         self.stock = []
 
     def __hash__(self):
+        """Uniquely identify a client"""
         return hash((self.name, self.phone_number))
 
     def add_to_stock(self, items: list[FoodItem]) -> None:
-        print("added to clients stock")
         for item in items:
             self.stock.append(item)
 
@@ -123,66 +110,21 @@ class Client:
             pizza_name, self, is_delivery=self.is_delivery
         )
         if not self.is_delivery:
-            food = self.pickup()
+            food = self._pickup()
             self.add_to_stock(food)
 
-    def pickup(self):
+    @log("Picking up took {:.2f} seconds")
+    @add_latency
+    def _pickup(self):
+        print("🏎️  Picking up", end=" ")
         food = self.restaurant.pickup(self)
         return food
 
 
-class DictAssortment(UserDict):
-    def __init__(self):
-        super().__init__()
-
-    def __getitem__(self, item: str):
-        """To accept pizza name indifferent of register"""
-        return super().__getitem__(item.lower())
-
-    def __setitem__(self, key, value):
-        return super().__setitem__(key.lower(), value)
-
-    def __contains__(self, item: str):
-        return super().__contains__(item.lower())
-
-
-# move it to different module maybe
-assortment = DictAssortment()
-
-
-def pizza_to_assortment(cls):
-    assortment[cls.__name__] = cls
-    return cls
-
-
-@pizza_to_assortment
-class Margherita(PizzaRecipe):
-    recipe = ["tomato sauce", "mozzarella", "tomatoes"]
-    emoji = "🧀"
-
-
-@pizza_to_assortment
-class Pepperoni(PizzaRecipe):
-    recipe = ["tomato sauce", "mozzarella", "pepperoni"]
-    emoji = "🍕"
-
-
-@pizza_to_assortment
-class Hawaiian(PizzaRecipe):
-    recipe = ["tomato sauce", "mozzarella", "chicken", "pineapples"]
-    emoji = "🍍"
-
-
-menu_str = "\n".join(
-    f"- {v.name} {v.emoji} : {v.clean_recipe}" for v in assortment.values()
-)
-
-
 if __name__ == "__main__":
-    # cli()
     restaurant = Restaurant(assortment)
     a = Pepperoni()
-    client = Client(restaurant=restaurant, is_delivery=True)
+    client = Client(restaurant=restaurant, is_delivery=False)
     client.order("Pepperoni")
     client.order("Pepperoni")
     print(client.stock)
